create Avishkar card
